[ut] reduce training step in test_dipole to reduce rtol
It only affects the test pass rate. 

```
"test_dipole.py failed randomly, since the relative error somehow is too big":

Yes, but look at how that is calculated. 
This is the difference between D_RMSE and calculating it manually. 
That has to be because the numbers are so small.
Otherwise, it cannot be wrong.
The model has a report D_RMSE of 6E-8.

You need a big relative error and a small absolute error.
are you sure you want to loose more rtol to fit the result?

No, I think if we train for very little steps it should be better.
The prediction error should be larger avoiding this problem.
```